separates the RapidStream route case
